Rename dev profile to kibana (#52)
We expect that users use this project for development purposes only. For
production usage they should use our helm charts. However, a section was
added to the README about development, which is likely confusing. This
section was about using Kibana to inspect, explore or troubleshoot the
the camunda platform data in Elasticsearch.

This commit renames the profile to `kibana` to better reflect the usage.
It also improves the README to better convey what this profile can be
used for. Lastly, it updates the note about index patterns, as this was
very specific to 1 of the use cases that kibana can be used for.